pyproject.toml
[project]
name = "live-speech-to-text-gcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-cloud-speech>=2.36.0",
    "google-cloud-storage>=3.8.0",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.2.1",
]
